Extract light environment model
